refactor: move SidebarOrgSection types and restore focus-visible
