--- docs/docs/auto-docs/types/SidebarOrgSection/interface/interfaces/ISidebarOrgSectionProps.md
@@ -1,41 +1,0 @@
-[Admin Docs](/)
-
-***
-
-# Interface: ISidebarOrgSectionProps
-
-Defined in: [src/types/SidebarOrgSection/interface.ts:9](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/SidebarOrgSection/interface.ts#L9)
-
-Interface for SidebarOrgSection component props.
-
- ISidebarOrgSectionProps
-
-## Properties
-
-### hideDrawer
-
-> **hideDrawer**: `boolean`
-
-Defined in: [src/types/SidebarOrgSection/interface.ts:11](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/SidebarOrgSection/interface.ts#L11)
-
-Whether the drawer is hidden/collapsed
-
-***
-
-### isProfilePage?
-
-> `optional` **isProfilePage**: `boolean`
-
-Defined in: [src/types/SidebarOrgSection/interface.ts:12](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/SidebarOrgSection/interface.ts#L12)
-
-Whether current page is the profile page
-
-***
-
-### orgId
-
-> **orgId**: `string`
-
-Defined in: [src/types/SidebarOrgSection/interface.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/SidebarOrgSection/interface.ts#L10)
-
-Organization ID to fetch and display
