+++ docs/docs/auto-docs/types/shared-components/SidebarOrgSection/interface/interfaces/ISidebarOrgSectionProps.md
@@ -1,0 +1,39 @@
+[Admin Docs](/)
+
+***
+
+# Interface: ISidebarOrgSectionProps
+
+Defined in: [src/types/shared-components/SidebarOrgSection/interface.ts:4](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarOrgSection/interface.ts#L4)
+
+Props for the SidebarOrgSection component.
+
+## Properties
+
+### hideDrawer
+
+> **hideDrawer**: `boolean`
+
+Defined in: [src/types/shared-components/SidebarOrgSection/interface.ts:8](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarOrgSection/interface.ts#L8)
+
+Whether the drawer is hidden/collapsed.
+
+***
+
+### isProfilePage?
+
+> `optional` **isProfilePage**: `boolean`
+
+Defined in: [src/types/shared-components/SidebarOrgSection/interface.ts:10](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarOrgSection/interface.ts#L10)
+
+Whether current page is the profile page.
+
+***
+
+### orgId
+
+> **orgId**: `string`
+
+Defined in: [src/types/shared-components/SidebarOrgSection/interface.ts:6](https://github.com/PalisadoesFoundation/talawa-admin/blob/main/src/types/shared-components/SidebarOrgSection/interface.ts#L6)
+
+Organization ID to fetch and display.
